+++ streets_utils/streets_messages/include/sensor_data_sharing_msg/angular_velocity_confidence.hpp
