+++ streets_utils/streets_messages/include/sensor_data_sharing_msg/equipement_type.hpp
