+++ streets_utils/streets_messages/include/sensor_data_sharing_msg/vehicle/attitude.hpp
@@ -13,7 +13,6 @@
 // limitations under the License.
 #pragma once
 
-#include <stdlib.h>
 
 namespace streets_utils::messages{
     struct attitude{
