+++ streets_utils/streets_messages/include/sensor_data_sharing_msg/vru/animal_propelled_type.hpp
