+++ streets_utils/streets_messages/include/sensor_data_sharing_msg/vru/motorized_propelled_type.hpp
