+++ streets_utils/streets_messages/include/serializers/sensor_data_sharing_msg_json_serializer.hpp
@@ -5,7 +5,7 @@
 #include <rapidjson/document.h>
 #include <rapidjson/writer.h>
 #include <rapidjson/stringbuffer.h>
-#include "messages/sensor_data_sharing_msg/sensor_data_sharing_msg.hpp"
+#include "sensor_data_sharing_msg/sensor_data_sharing_msg.hpp"
 #include <streets_utils/json_utils_lib/json_utils_exception.hpp>
