+++ streets_utils/streets_messages/test/personal_device_user_type.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <messages/sensor_data_sharing_msg/vru/personal_device_user_type.hpp>
+#include <sensor_data_sharing_msg/vru/personal_device_user_type.hpp>
 
 using namespace streets_utils::messages;
